Perform various DB migration fixes and improvements

+++ application/migrations/010_add_time_format_setting.php
@@ -17,10 +17,13 @@ class Migration_Add_time_format_setting extends EA_Migration {
      */
     public function up()
     {
-        $this->db->insert('settings', [
-            'name' => 'time_format',
-            'value' => 'regular'
-        ]);
+        if ( ! $this->db->get_where('settings', ['name' => 'time_format'])->num_rows())
+        {
+            $this->db->insert('settings', [
+                'name' => 'time_format',
+                'value' => 'regular'
+            ]);
+        }
     }
 
     /**
@@ -28,6 +31,9 @@ class Migration_Add_time_format_setting extends EA_Migration {
      */
     public function down()
     {
-        $this->db->delete('settings', ['name' => 'time_format']);
+        if ($this->db->get_where('settings', ['name' => 'time_format'])->num_rows())
+        {
+            $this->db->delete('settings', ['name' => 'time_format']);
+        }
     }
 }
