+++ application/migrations/017_add_api_token_setting.php
@@ -19,10 +19,14 @@ class Migration_Add_api_token_setting extends EA_Migration {
      */
     public function up()
     {
-        $this->db->insert('settings', [
-            'name' => 'api_token',
-            'value' => ''
-        ]);
+        if ( ! $this->db->get_where('settings', ['name' => 'api_token'])->num_rows())
+        {
+            $this->db->insert('settings', [
+                'name' => 'api_token',
+                'value' => ''
+            ]);
+        }
+
     }
 
     /**
@@ -32,6 +36,9 @@ class Migration_Add_api_token_setting extends EA_Migration {
      */
     public function down()
     {
-        $this->db->delete('settings', ['name' => 'api_token']);
+        if ($this->db->get_where('settings', ['name' => 'api_token'])->num_rows())
+        {
+            $this->db->delete('settings', ['name' => 'api_token']);
+        }
     }
 }
